fix validation

+++ src/app/modules/Order/order.services.ts
@@ -2,6 +2,12 @@ import prisma from "../../utils/prisma";
 import { OrderType } from "./order.interface";
 
 const createOrderDB = async (payload: OrderType) => {
+    const exitingId = await prisma.product.findUnique({
+    where: { id:payload.productId },
+  });
+  if (!exitingId) {
+    throw new Error(" Product id Not Fund!");
+  }
   return prisma.order.create({ data: payload });
 };
 
@@ -10,6 +16,12 @@ const getOrderDB = async () => {
 };
 
 const deleteOrderDB = async (orderId: string) => {
+  const exitingId = await prisma.order.findUnique({
+    where: { id: orderId },
+  });
+  if (!exitingId) {
+    throw new Error("Id Not Fund!");
+  }
   return prisma.order.delete({ where: { id: orderId } });
 };
